v2.3.5.1

+++ Projectiles/SeekerKnifeProj.cs
@@ -21,7 +21,7 @@ namespace StormDiversMod.Projectiles
 
         public override void SetStaticDefaults()
         {
-            DisplayName.SetDefault("Seeeker Knife");
+            DisplayName.SetDefault("Seeker Knife");
             ProjectileID.Sets.TrailingMode[Projectile.type] = 2;
             ProjectileID.Sets.TrailCacheLength[Projectile.type] = 6;
             ProjectileID.Sets.CultistIsResistantTo[Projectile.type] = true;
@@ -126,8 +126,10 @@ namespace StormDiversMod.Projectiles
         }
         public override bool OnTileCollide(Vector2 oldVelocity)
         {
-
-            Projectile.Kill();
+            if (hometime > 5)
+            {
+                Projectile.Kill();
+            }
             return true;
         }
